+++ src/App.jsx
@@ -1,4 +1,4 @@
-import Footer from "./components/Footer";
+import Footer from "./components/footer/Footer";
 import Header from "./components/Header";
 import LandingContent from "./components/LandingContent";
